refactor narsese module
