+++ pynars/Global.py
@@ -1,14 +1,7 @@
 import inspect
+from pynars.Narsese import Global as _Global
 
-time = 0
-_input_id = 0
-def get_input_id():
-    global _input_id
-    input_id = _input_id
-    _input_id += 1
-    return input_id
-
-class States:
+class StatesCls:
     task = None
     belief = None
     concept = None
@@ -19,9 +12,12 @@ class States:
         ''''''
         for name in vars(cls):
             attr = getattr(cls, name)
-            if inspect.ismethod(attr): continue
-            if name.startswith('_'): continue
-            if isinstance(attr, property): continue
+            if inspect.ismethod(attr):
+                continue
+            if name.startswith('_'):
+                continue
+            if isinstance(attr, property):
+                continue
             # if isinstance(attr, classmethod): continue
             setattr(cls, name, None)
 
@@ -29,7 +25,7 @@ class States:
     def record_premises(cls, task=None, belief=None):
         cls.task = task
         cls.belief = belief
-    
+
     @classmethod
     def record_concept(cls, concept=None):
         cls.concept = concept
@@ -37,11 +33,22 @@ class States:
     @classmethod
     def record_rules(cls, rules=None):
         cls.rules = rules
-    
+
     @property
     def time(self):
-        return time
-    
+        return _Global.time
+
+    @time.setter
+    def time(self, value):
+        _Global.time = value
+
+    @staticmethod
+    def get_input_id():
+        return _Global.get_input_id()
+
     @classmethod
     def __repr__(cls):
-        return f'<States: time={time}\n\tconcept: {cls.concept}\n\ttask: {cls.task}\n\tbelief: {cls.belief}\n\trules: {cls.rules}\n>.'
+        return f'<States: time={cls.time}\n\tconcept: {cls.concept}\n\ttask: {cls.task}\n\tbelief: {cls.belief}\n\trules: {cls.rules}\n>.'
+
+
+Global = StatesCls()
